Conflicts:
	components/tools/OmeroWeb/omeroweb/urls.py

+++ components/insight/SRC/org/openmicroscopy/shoola/env/data/OmeroMetadataServiceImpl.java
@@ -181,13 +181,11 @@ class OmeroMetadataServiceImpl
 		throws DSOutOfServiceException, DSAccessException 
 	{
 		if (annotation == null || object == null) return;
-		ExperimenterData exp = getUserDetails();
-		if (exp == null) return;
 		IObject ho = gateway.findIObject(ctx, annotation.asIObject());
-		if (ho == null) return;
+		if (ho == null || !gateway.canDelete(ho)) return;
 		IObject link = gateway.findAnnotationLink(ctx, object.getClass(),
-				       object.getId(), ho.getId().getValue(), exp.getId());
-		if (ho != null && link != null) {
+				       object.getId(), ho.getId().getValue(), -1);
+		if (link != null) {
 			try {
 				gateway.deleteObject(ctx, link);
 			} catch (Exception e) {
